move IStyles interface from styles.tsx to the new models folder
